requirements.txt
Flask==3.1.3
flask-cors==6.0.2
torch==2.11.0
asteroid==0.7.0
asteroid-filterbanks==0.4.0
huggingface_hub==1.8.0
librosa==0.11.0
soundfile==0.13.1
numpy==2.4.4
scipy==1.17.1
mediapipe==0.10.33
opencv-python==4.13.0.92
matplotlib==3.10.8
python-dotenv==1.2.2
